requirements.txt
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch>=0.7.1
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.65.0
pandas>=1.5.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
albumentations>=1.3.0
tensorboard>=2.12.0
PyYAML>=6.0
requests>=2.28.0
urllib3>=1.26.0
certifi>=2022.12.7
charset-normalizer>=2.0.0
idna>=3.4
six>=1.16.0
packaging>=23.0
typing-extensions>=4.5.0